🔧 Refactor code to add places and markers to the map
This commit refactors the code in `new.vue` to add functionality for adding places and markers to the map. It introduces a new `addPlace` function that adds a place to the list of places and creates a marker on the map. The marker is associated with an info window that displays information about the place when clicked. Additionally, a new `focusPlace` function is added to open the info window of a specific place and close all other info windows. The map is now bounded to fit all markers using the `boundMap` function.